// client/src/components/cv-builder/AIAssistance/enums.ts
export enum Prompts {
  fixSpelling = "fixSpelling",
  rephrase = "rephrase",
  generate = "generate",
}
export enum States {
  Initial = "Initial",
  Loading = "Loading",
  Response = "Response",
}

export enum Namespaces {
  experience = "experience",
  skills = "skills",
  summary = "summary",
}